refactor: break out analyzing class to own class.
